feat: restructure shared utils and update deploy command
- Move shared libraries to a dedicated folder
- Add command to build the shared libraries
- Re-write deploy command for improved functionality
- Update README.md with new instructions
