+++ functions/notify_token_service/requirements.txt
@@ -1,6 +1,4 @@
 requests==2.26.0
 functions-framework==3.*
 flask==3.0.0
-sqlalchemy==2.0.23
-cloud-sql-python-connector
-psycopg2
+dist/shared_lib-0.0.1.tar.gz
