+++ functions/weekly_game_notify/requirements.txt
@@ -1,5 +1,3 @@
 requests==2.26.0
 functions-framework==3.*
-google-api-python-client
-google-auth-httplib2
-google-auth-oauthlib
+dist/shared_lib-0.0.1.tar.gz
