+++ requirements.txt
@@ -1,5 +1,6 @@
 isort==5.13.2
 black==24.4.2
+setuptools
 
 requests==2.26.0
 functions-framework==3.*
@@ -10,3 +11,5 @@ psycopg2
 google-api-python-client
 google-auth-httplib2
 google-auth-oauthlib
+
+shared_lib/dist/shared_lib-0.0.1.tar.gz
