+++ shared_lib/README.md
@@ -1,0 +1,3 @@
+# Shared Library
+
+This is a shared library for NTUBTOB-management-system.
