+++ shared_lib/shared_module/games_crawler_client.py
@@ -1,7 +1,8 @@
 from dataclasses import dataclass
 
 import requests
-from gcloud import get_id_token
+
+from .gcloud import get_id_token
 
 
 @dataclass
