+++ shared_lib/shared_module/gcloud.py
