+++ shared_lib/shared_module/message.py
